Refactor libazureinit-kvp: split into error/store modules, improve coverage
- Split kvp_pool.rs into error.rs (KvpError) and store.rs (KvpStore trait + KvpPoolStore)
- lib.rs reduced to module declarations and re-exports
- Replace infallible dump_json map_err with expect
- Inline get_uptime into boot_time
- Add permission-denied, stale-clear, and missing-key-delete tests
- 98.5% line coverage; remaining gaps are untestable kernel-level fcntl error paths
