+++ libazureinit-kvp/src/store.rs
@@ -1,0 +1,2105 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+//! Unified KVP pool file backend for Hyper-V and Azure guests.
+//!
+//! The on-disk pool format is fixed-width and identical across
+//! environments:
+//! - key field: 512 bytes
+//! - value field: 2048 bytes
+//! - record size: 2560 bytes
+//!
+//! [`PoolMode`] selects which size limits are enforced on writes:
+//! - [`Safe`](PoolMode::Safe): key <= 254 bytes,
+//!   value <= 1022 bytes (recommended for Linux kernel compatibility)
+//! - [`Unsafe`](PoolMode::Unsafe): key <= 512 bytes, value <= 2048 bytes
+//!
+//! ## Reference
+//! - [Hyper-V Data Exchange Service (KVP)](https://learn.microsoft.com/en-us/virtualization/hyper-v-on-windows/reference/integration-services#hyper-v-data-exchange-service-kvp)
+
+use std::collections::{HashMap, HashSet};
+use std::fs::{File, OpenOptions};
+use std::io::{self, ErrorKind, Read, Seek, Write};
+use std::os::unix::fs::MetadataExt;
+use std::os::unix::io::AsRawFd;
+use std::path::{Path, PathBuf};
+use std::time::{SystemTime, UNIX_EPOCH};
+
+use sysinfo::System;
+
+use crate::KvpError;
+/// Default base directory for KVP pool files.
+const DEFAULT_KVP_DIR: &str = "/var/lib/hyperv";
+
+const WIRE_MAX_KEY_BYTES: usize = 512;
+const WIRE_MAX_VALUE_BYTES: usize = 2048;
+const SAFE_MAX_KEY_BYTES: usize = 254;
+const SAFE_MAX_VALUE_BYTES: usize = 1022;
+
+/// Maximum number of unique keys allowed in the pool.
+const MAX_UNIQUE_KEYS: usize = 1024;
+
+const RECORD_SIZE: usize = WIRE_MAX_KEY_BYTES + WIRE_MAX_VALUE_BYTES;
+/// Hyper-V KVP pool indices.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+#[repr(u8)]
+pub enum KvpPool {
+    /// Host-to-guest data pushed by the host admin (`.kvp_pool_0`).
+    External = 0,
+    /// Guest-to-host data; cloud-init / azure-init write here (`.kvp_pool_1`).
+    Guest = 1,
+    /// Guest intrinsics generated by the daemon (`.kvp_pool_2`).
+    Auto = 2,
+    /// Host-originated data describing the host (`.kvp_pool_3`).
+    AutoExternal = 3,
+    /// Undocumented; no pool file exists (`.kvp_pool_4`).
+    AutoInternal = 4,
+}
+
+impl KvpPool {
+    /// File name of the pool file, e.g. `.kvp_pool_1`.
+    pub fn file_name(self) -> &'static str {
+        match self {
+            Self::External => ".kvp_pool_0",
+            Self::Guest => ".kvp_pool_1",
+            Self::Auto => ".kvp_pool_2",
+            Self::AutoExternal => ".kvp_pool_3",
+            Self::AutoInternal => ".kvp_pool_4",
+        }
+    }
+
+    /// Default base directory for KVP pool files.
+    pub fn default_dir() -> &'static Path {
+        Path::new(DEFAULT_KVP_DIR)
+    }
+
+    /// The conventional on-disk path (`/var/lib/hyperv/<pool_file>`).
+    pub fn default_path(self) -> PathBuf {
+        Self::default_dir().join(self.file_name())
+    }
+}
+
+/// Policy mode controlling key/value size limits for writes.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum PoolMode {
+    /// Conservative limits for Linux kernel compatibility
+    /// (key <= 254 bytes, value <= 1022 bytes).
+    Safe,
+    /// Full Hyper-V wire-format limits
+    /// (key <= 512 bytes, value <= 2048 bytes).
+    Unsafe,
+}
+
+impl PoolMode {
+    fn max_key_size(self) -> usize {
+        match self {
+            Self::Safe => SAFE_MAX_KEY_BYTES,
+            Self::Unsafe => WIRE_MAX_KEY_BYTES,
+        }
+    }
+
+    fn max_value_size(self) -> usize {
+        match self {
+            Self::Safe => SAFE_MAX_VALUE_BYTES,
+            Self::Unsafe => WIRE_MAX_VALUE_BYTES,
+        }
+    }
+}
+
+/// Key-value store with Hyper-V KVP semantics.
+pub trait KvpStore: Send + Sync {
+    /// Maximum key size in bytes for writes.
+    fn max_key_size(&self) -> usize;
+
+    /// Maximum value size in bytes for writes.
+    fn max_value_size(&self) -> usize;
+
+    /// Insert a new key-value pair or update an existing key's value.
+    fn insert(&self, key: &str, value: &str) -> Result<(), KvpError>;
+
+    /// Append a key-value pair without checking for an existing key.
+    fn append(&self, key: &str, value: &str) -> Result<(), KvpError>;
+
+    /// Read the value for a key. Returns `Ok(None)` when absent.
+    ///
+    /// If multiple records share the same key (e.g. via
+    /// [`append`](Self::append)), the last (most recent) match wins.
+    fn read(&self, key: &str) -> Result<Option<String>, KvpError>;
+
+    /// Remove all records matching the key. Returns `true` if at least
+    /// one record was present.
+    fn delete(&self, key: &str) -> Result<bool, KvpError>;
+
+    /// Remove all entries from the store.
+    fn clear(&self) -> Result<(), KvpError>;
+
+    /// Return all key-value pairs (deduplicated, last-write-wins).
+    fn entries(&self) -> Result<HashMap<String, String>, KvpError>;
+
+    /// Return all key-value records in on-disk order, including
+    /// duplicates from [`append`](Self::append) calls.
+    fn dump(&self) -> Result<Vec<(String, String)>, KvpError>;
+
+    /// Return the number of records in the store.
+    ///
+    /// This counts on-disk records, not unique keys. If
+    /// [`append`](Self::append) was used to write duplicate keys, this
+    /// may exceed the number of unique keys returned by
+    /// [`entries`](Self::entries).
+    fn len(&self) -> Result<usize, KvpError>;
+
+    /// Return whether the store is empty.
+    fn is_empty(&self) -> Result<bool, KvpError>;
+
+    /// Whether the store's data is stale (e.g. predates current boot).
+    fn is_stale(&self) -> Result<bool, KvpError>;
+}
+
+/// Unified KVP pool file store.
+#[derive(Clone, Debug)]
+pub struct KvpPoolStore {
+    pool: KvpPool,
+    path: PathBuf,
+    mode: PoolMode,
+}
+
+impl KvpPoolStore {
+    /// Open using the default Hyper-V directory
+    /// ([`/var/lib/hyperv`](KvpPool::default_dir)).
+    pub fn new(pool: KvpPool, mode: PoolMode) -> Result<Self, KvpError> {
+        Self::new_in(pool, KvpPool::default_dir(), mode)
+    }
+
+    /// Open using a custom directory (file name derived from `pool`).
+    pub fn new_in(
+        pool: KvpPool,
+        dir: impl AsRef<Path>,
+        mode: PoolMode,
+    ) -> Result<Self, KvpError> {
+        let path = dir.as_ref().join(pool.file_name());
+        Ok(Self { pool, path, mode })
+    }
+
+    /// The pool this store was created for.
+    pub fn pool(&self) -> KvpPool {
+        self.pool
+    }
+
+    /// Return a reference to the pool file path.
+    pub fn path(&self) -> &Path {
+        &self.path
+    }
+
+    /// The policy mode this store was created with.
+    pub fn mode(&self) -> PoolMode {
+        self.mode
+    }
+
+    /// Clear the pool file if it contains data from a previous boot.
+    pub fn clear_if_stale(&self) -> Result<(), KvpError> {
+        if self.is_stale()? {
+            self.clear()?;
+        }
+        Ok(())
+    }
+
+    /// Dump all entries (deduplicated) to a JSON file at the given path.
+    pub fn dump_json(&self, path: &Path) -> Result<(), KvpError> {
+        let entries = self.entries()?;
+        let json = serde_json::to_string_pretty(&entries)
+            .expect("HashMap<String, String> serialization is infallible");
+        std::fs::write(path, json)?;
+        Ok(())
+    }
+
+    fn validate_key(&self, key: &str) -> Result<(), KvpError> {
+        if key.is_empty() {
+            return Err(KvpError::EmptyKey);
+        }
+        if key.as_bytes().contains(&0) {
+            return Err(KvpError::KeyContainsNull);
+        }
+        let actual = key.len();
+        let max = self.mode.max_key_size();
+        if actual > max {
+            return Err(KvpError::KeyTooLarge { max, actual });
+        }
+        Ok(())
+    }
+
+    /// Looser validation for reads: accepts keys up to the full
+    /// wire-format maximum regardless of [`PoolMode`].
+    fn validate_key_for_read(key: &str) -> Result<(), KvpError> {
+        if key.is_empty() {
+            return Err(KvpError::EmptyKey);
+        }
+        if key.as_bytes().contains(&0) {
+            return Err(KvpError::KeyContainsNull);
+        }
+        let actual = key.len();
+        if actual > WIRE_MAX_KEY_BYTES {
+            return Err(KvpError::KeyTooLarge {
+                max: WIRE_MAX_KEY_BYTES,
+                actual,
+            });
+        }
+        Ok(())
+    }
+
+    fn validate_value(&self, value: &str) -> Result<(), KvpError> {
+        let actual = value.len();
+        let max = self.mode.max_value_size();
+        if actual > max {
+            return Err(KvpError::ValueTooLarge { max, actual });
+        }
+        Ok(())
+    }
+
+    fn boot_time() -> Result<i64, KvpError> {
+        let now = SystemTime::now()
+            .duration_since(UNIX_EPOCH)
+            .map_err(|e| io::Error::other(format!("clock error: {e}")))?
+            .as_secs();
+        Ok(now.saturating_sub(System::uptime()) as i64)
+    }
+
+    /// Test-only variant of [`is_stale`](KvpStore::is_stale) that accepts
+    /// an explicit boot time for deterministic assertions.
+    #[cfg(test)]
+    fn is_stale_at_boot(&self, boot_time: i64) -> Result<bool, KvpError> {
+        let metadata = match std::fs::metadata(&self.path) {
+            Ok(m) => m,
+            Err(ref e) if e.kind() == ErrorKind::NotFound => return Ok(false),
+            Err(e) => return Err(e.into()),
+        };
+        Ok(metadata.mtime() < boot_time)
+    }
+
+    fn open_for_read(&self) -> io::Result<File> {
+        OpenOptions::new().read(true).open(&self.path)
+    }
+
+    fn open_for_read_write(&self) -> io::Result<File> {
+        OpenOptions::new().read(true).write(true).open(&self.path)
+    }
+
+    fn open_for_read_write_create(&self) -> io::Result<File> {
+        OpenOptions::new()
+            .read(true)
+            .write(true)
+            .create(true)
+            .truncate(false)
+            .open(&self.path)
+    }
+
+    /// Create a read-only iterator over all records.
+    fn iter(&self) -> Result<KvpPoolIter, KvpError> {
+        let file = self.open_for_read()?;
+        KvpPoolIter::new(file, false)
+    }
+
+    /// Create a read-write iterator over all records.
+    fn iter_mut(&self) -> Result<KvpPoolIter, KvpError> {
+        let file = self.open_for_read_write_create()?;
+        KvpPoolIter::new(file, true)
+    }
+}
+
+impl KvpStore for KvpPoolStore {
+    fn max_key_size(&self) -> usize {
+        self.mode.max_key_size()
+    }
+
+    fn max_value_size(&self) -> usize {
+        self.mode.max_value_size()
+    }
+
+    fn insert(&self, key: &str, value: &str) -> Result<(), KvpError> {
+        self.validate_key(key)?;
+        self.validate_value(value)?;
+
+        let mut iter = self.iter_mut()?;
+        let mut found = false;
+        let mut unique_keys = HashSet::new();
+
+        while let Some(record) = iter.next() {
+            let (k, _) = record?;
+            let is_target = k == key;
+            unique_keys.insert(k);
+            if is_target {
+                if !found {
+                    iter.overwrite_current_value(value)?;
+                    found = true;
+                } else {
+                    iter.remove_current()?;
+                }
+            }
+        }
+
+        if !found {
+            if unique_keys.len() >= MAX_UNIQUE_KEYS {
+                return Err(KvpError::MaxUniqueKeysExceeded {
+                    max: MAX_UNIQUE_KEYS,
+                });
+            }
+            iter.append(key, value)?;
+        }
+
+        iter.flush()?;
+        Ok(())
+    }
+
+    fn append(&self, key: &str, value: &str) -> Result<(), KvpError> {
+        self.validate_key(key)?;
+        self.validate_value(value)?;
+
+        let mut iter = self.iter_mut()?;
+        iter.append(key, value)?;
+        iter.flush()?;
+        Ok(())
+    }
+
+    fn read(&self, key: &str) -> Result<Option<String>, KvpError> {
+        Self::validate_key_for_read(key)?;
+
+        let iter = match self.iter() {
+            Ok(it) => it,
+            Err(KvpError::Io(e)) if e.kind() == ErrorKind::NotFound => {
+                return Ok(None);
+            }
+            Err(e) => return Err(e),
+        };
+
+        let mut result = None;
+        for record in iter {
+            let (k, v) = record?;
+            if k == key {
+                result = Some(v);
+            }
+        }
+
+        Ok(result)
+    }
+
+    fn delete(&self, key: &str) -> Result<bool, KvpError> {
+        Self::validate_key_for_read(key)?;
+
+        let file = match self.open_for_read_write() {
+            Ok(f) => f,
+            Err(ref e) if e.kind() == ErrorKind::NotFound => {
+                return Ok(false);
+            }
+            Err(e) => return Err(e.into()),
+        };
+
+        let mut iter = KvpPoolIter::new(file, true)?;
+        let mut found = false;
+
+        while let Some(record) = iter.next() {
+            let (k, _) = record?;
+            if k == key {
+                iter.remove_current()?;
+                found = true;
+            }
+        }
+
+        if found {
+            iter.flush()?;
+        }
+        Ok(found)
+    }
+
+    fn clear(&self) -> Result<(), KvpError> {
+        let file =
+            match OpenOptions::new().read(true).write(true).open(&self.path) {
+                Ok(f) => f,
+                Err(ref e) if e.kind() == ErrorKind::NotFound => return Ok(()),
+                Err(e) => return Err(e.into()),
+            };
+
+        fcntl_lock_exclusive(&file).map_err(|e| {
+            io::Error::other(format!("failed to lock KVP file: {e}"))
+        })?;
+        let result = file.set_len(0).map_err(KvpError::from);
+        let _ = fcntl_unlock(&file);
+        result
+    }
+
+    fn entries(&self) -> Result<HashMap<String, String>, KvpError> {
+        let iter = match self.iter() {
+            Ok(it) => it,
+            Err(KvpError::Io(e)) if e.kind() == ErrorKind::NotFound => {
+                return Ok(HashMap::new());
+            }
+            Err(e) => return Err(e),
+        };
+
+        let mut map = HashMap::with_capacity(iter.record_count());
+        for record in iter {
+            let (k, v) = record?;
+            map.insert(k, v);
+        }
+        Ok(map)
+    }
+
+    fn dump(&self) -> Result<Vec<(String, String)>, KvpError> {
+        let iter = match self.iter() {
+            Ok(it) => it,
+            Err(KvpError::Io(e)) if e.kind() == ErrorKind::NotFound => {
+                return Ok(Vec::new());
+            }
+            Err(e) => return Err(e),
+        };
+
+        let mut records = Vec::with_capacity(iter.record_count());
+        for record in iter {
+            records.push(record?);
+        }
+        Ok(records)
+    }
+
+    fn len(&self) -> Result<usize, KvpError> {
+        match std::fs::metadata(&self.path) {
+            Ok(m) => Ok(m.len() as usize / RECORD_SIZE),
+            Err(ref e) if e.kind() == ErrorKind::NotFound => Ok(0),
+            Err(e) => Err(e.into()),
+        }
+    }
+
+    fn is_empty(&self) -> Result<bool, KvpError> {
+        Ok(self.len()? == 0)
+    }
+
+    fn is_stale(&self) -> Result<bool, KvpError> {
+        let metadata = match std::fs::metadata(&self.path) {
+            Ok(m) => m,
+            Err(ref e) if e.kind() == ErrorKind::NotFound => return Ok(false),
+            Err(e) => return Err(e.into()),
+        };
+        let boot = Self::boot_time()?;
+        Ok(metadata.mtime() < boot)
+    }
+}
+
+/// Acquire an exclusive (write) lock on the entire file.
+///
+/// Uses `fcntl(F_OFD_SETLKW)` — open file description locks that are
+/// per-FD (safe for multi-threaded use) yet conflict with traditional
+/// `fcntl` record locks used by `hv_kvp_daemon.c` and cloud-init.
+fn fcntl_lock_exclusive(file: &File) -> io::Result<()> {
+    let fl = libc::flock {
+        l_type: libc::F_WRLCK as libc::c_short,
+        l_whence: libc::SEEK_SET as libc::c_short,
+        l_start: 0,
+        l_len: 0,
+        l_pid: 0,
+    };
+    if unsafe { libc::fcntl(file.as_raw_fd(), libc::F_OFD_SETLKW, &fl) } == -1 {
+        return Err(io::Error::last_os_error());
+    }
+    Ok(())
+}
+
+/// Acquire a shared (read) lock on the entire file.
+fn fcntl_lock_shared(file: &File) -> io::Result<()> {
+    let fl = libc::flock {
+        l_type: libc::F_RDLCK as libc::c_short,
+        l_whence: libc::SEEK_SET as libc::c_short,
+        l_start: 0,
+        l_len: 0,
+        l_pid: 0,
+    };
+    if unsafe { libc::fcntl(file.as_raw_fd(), libc::F_OFD_SETLKW, &fl) } == -1 {
+        return Err(io::Error::last_os_error());
+    }
+    Ok(())
+}
+
+/// Release a lock on the entire file.
+fn fcntl_unlock(file: &File) -> io::Result<()> {
+    let fl = libc::flock {
+        l_type: libc::F_UNLCK as libc::c_short,
+        l_whence: libc::SEEK_SET as libc::c_short,
+        l_start: 0,
+        l_len: 0,
+        l_pid: 0,
+    };
+    if unsafe { libc::fcntl(file.as_raw_fd(), libc::F_OFD_SETLK, &fl) } == -1 {
+        return Err(io::Error::last_os_error());
+    }
+    Ok(())
+}
+
+pub(crate) fn encode_record(key: &str, value: &str) -> Vec<u8> {
+    let mut buf = vec![0u8; RECORD_SIZE];
+
+    let key_bytes = key.as_bytes();
+    let key_len = key_bytes.len().min(WIRE_MAX_KEY_BYTES);
+    buf[..key_len].copy_from_slice(&key_bytes[..key_len]);
+
+    let value_bytes = value.as_bytes();
+    let value_len = value_bytes.len().min(WIRE_MAX_VALUE_BYTES);
+    buf[WIRE_MAX_KEY_BYTES..WIRE_MAX_KEY_BYTES + value_len]
+        .copy_from_slice(&value_bytes[..value_len]);
+
+    buf
+}
+
+pub(crate) fn decode_record(data: &[u8]) -> io::Result<(String, String)> {
+    if data.len() != RECORD_SIZE {
+        return Err(io::Error::other(format!(
+            "record size mismatch: expected {RECORD_SIZE}, got {}",
+            data.len()
+        )));
+    }
+
+    let (key_bytes, value_bytes) = data.split_at(WIRE_MAX_KEY_BYTES);
+
+    let key = std::str::from_utf8(key_bytes)
+        .map_err(|e| io::Error::new(ErrorKind::InvalidData, e))?
+        .trim_end_matches('\0')
+        .to_string();
+    let value = std::str::from_utf8(value_bytes)
+        .map_err(|e| io::Error::new(ErrorKind::InvalidData, e))?
+        .trim_end_matches('\0')
+        .to_string();
+
+    Ok((key, value))
+}
+
+/// A record-at-a-time iterator over a KVP pool file.
+///
+/// Uses file I/O with seek semantics for memory-efficient iteration
+/// without loading all records into memory. Supports in-place value
+/// overwrites and record removal for the last-yielded record.
+///
+/// The underlying file is locked for the lifetime of the iterator;
+/// the lock is released on drop.
+#[derive(Debug)]
+struct KvpPoolIter {
+    file: File,
+    record_count: usize,
+    current_index: usize,
+}
+
+impl KvpPoolIter {
+    fn new(mut file: File, lock_exclusive: bool) -> Result<Self, KvpError> {
+        let lock_result = if lock_exclusive {
+            fcntl_lock_exclusive(&file)
+        } else {
+            fcntl_lock_shared(&file)
+        };
+        lock_result.map_err(|e| {
+            io::Error::other(format!("failed to lock KVP file: {e}"))
+        })?;
+
+        let len = file.metadata()?.len() as usize;
+        if len > 0 && !len.is_multiple_of(RECORD_SIZE) {
+            let _ = fcntl_unlock(&file);
+            return Err(io::Error::other(format!(
+                "file size ({len}) is not a multiple of record size \
+                 ({RECORD_SIZE})"
+            ))
+            .into());
+        }
+
+        let record_count = len / RECORD_SIZE;
+        file.seek(io::SeekFrom::Start(0))?;
+
+        Ok(Self {
+            file,
+            record_count,
+            current_index: 0,
+        })
+    }
+
+    /// The number of records in the file (updated by
+    /// [`append`](Self::append) and
+    /// [`remove_current`](Self::remove_current)).
+    fn record_count(&self) -> usize {
+        self.record_count
+    }
+
+    /// Overwrite the value field of the record last returned by
+    /// [`next()`](Iterator::next), zero-padding to fill the
+    /// fixed-width 2048-byte field.
+    ///
+    /// After the write the file position is at the start of the next
+    /// record, so iteration can continue normally.
+    ///
+    /// Returns `InvalidInput` if called before the first `next()`.
+    fn overwrite_current_value(&mut self, value: &str) -> io::Result<()> {
+        if self.current_index == 0 {
+            return Err(io::Error::new(
+                ErrorKind::InvalidInput,
+                "no record has been read yet",
+            ));
+        }
+
+        let record_start = (self.current_index - 1) as u64 * RECORD_SIZE as u64;
+        let value_offset = record_start + WIRE_MAX_KEY_BYTES as u64;
+
+        self.file.seek(io::SeekFrom::Start(value_offset))?;
+
+        let mut buf = [0u8; WIRE_MAX_VALUE_BYTES];
+        let bytes = value.as_bytes();
+        let len = bytes.len().min(WIRE_MAX_VALUE_BYTES);
+        buf[..len].copy_from_slice(&bytes[..len]);
+
+        self.file.write_all(&buf)?;
+        Ok(())
+    }
+
+    /// Append a new record at the end of the file, zero-padded to the
+    /// fixed-width wire format.
+    fn append(&mut self, key: &str, value: &str) -> io::Result<()> {
+        let pos = self.file.seek(io::SeekFrom::End(0))?;
+        debug_assert!(
+            (pos as usize).is_multiple_of(RECORD_SIZE),
+            "file position {pos} is not aligned to record size {RECORD_SIZE}"
+        );
+        self.file.write_all(&encode_record(key, value))?;
+        self.record_count += 1;
+        Ok(())
+    }
+
+    /// Remove the record last returned by [`next()`](Iterator::next)
+    /// by swapping it with the final record and truncating the file.
+    ///
+    /// After removal the file position rewinds so that the next call
+    /// to `next()` reads the record that was swapped into this slot
+    /// (unless the removed record was already the last one).
+    ///
+    /// Returns `InvalidInput` if called before the first `next()`.
+    fn remove_current(&mut self) -> io::Result<()> {
+        if self.current_index == 0 {
+            return Err(io::Error::new(
+                ErrorKind::InvalidInput,
+                "no record has been read yet",
+            ));
+        }
+
+        let delete_index = self.current_index - 1;
+        let last_index = self.record_count - 1;
+
+        if delete_index != last_index {
+            self.file.seek(io::SeekFrom::Start(
+                last_index as u64 * RECORD_SIZE as u64,
+            ))?;
+            let mut buf = [0u8; RECORD_SIZE];
+            self.file.read_exact(&mut buf)?;
+
+            self.file.seek(io::SeekFrom::Start(
+                delete_index as u64 * RECORD_SIZE as u64,
+            ))?;
+            self.file.write_all(&buf)?;
+
+            self.file.seek(io::SeekFrom::Start(
+                delete_index as u64 * RECORD_SIZE as u64,
+            ))?;
+            self.current_index = delete_index;
+        }
+
+        self.file.set_len(last_index as u64 * RECORD_SIZE as u64)?;
+        self.record_count -= 1;
+        Ok(())
+    }
+
+    /// Flush any buffered writes to the underlying file.
+    fn flush(&mut self) -> io::Result<()> {
+        self.file.flush()
+    }
+}
+
+impl Drop for KvpPoolIter {
+    fn drop(&mut self) {
+        let _ = fcntl_unlock(&self.file);
+    }
+}
+
+impl Iterator for KvpPoolIter {
+    type Item = io::Result<(String, String)>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.current_index >= self.record_count {
+            return None;
+        }
+        let mut buf = [0u8; RECORD_SIZE];
+        match self.file.read_exact(&mut buf) {
+            Ok(()) => {
+                self.current_index += 1;
+                Some(decode_record(&buf))
+            }
+            Err(e) => Some(Err(e)),
+        }
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        let remaining = self.record_count - self.current_index;
+        (remaining, Some(remaining))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::error::Error;
+    use std::os::unix::fs::PermissionsExt;
+    use tempfile::{NamedTempFile, TempDir};
+
+    fn safe_store(dir: &Path) -> KvpPoolStore {
+        KvpPoolStore::new_in(KvpPool::Guest, dir, PoolMode::Safe).unwrap()
+    }
+
+    fn unsafe_store(dir: &Path) -> KvpPoolStore {
+        KvpPoolStore::new_in(KvpPool::Guest, dir, PoolMode::Unsafe).unwrap()
+    }
+
+    /// Write `count` unique records directly to the pool file in a
+    /// single pass, bypassing the insert-per-key overhead.  Keys are
+    /// `k0`, `k1`, ... `k{count-1}` with value `"v"`.
+    fn prepopulate(store: &KvpPoolStore, count: usize) {
+        let mut file = File::create(store.path()).unwrap();
+        for i in 0..count {
+            file.write_all(&encode_record(&format!("k{i}"), "v"))
+                .unwrap();
+        }
+        file.flush().unwrap();
+    }
+
+    fn lock_dir(dir: &Path) {
+        std::fs::set_permissions(dir, std::fs::Permissions::from_mode(0o000))
+            .unwrap();
+    }
+
+    fn unlock_dir(dir: &Path) {
+        std::fs::set_permissions(dir, std::fs::Permissions::from_mode(0o755))
+            .unwrap();
+    }
+
+    fn set_mtime_to_epoch(path: &Path) {
+        let c_path =
+            std::ffi::CString::new(path.as_os_str().as_encoded_bytes())
+                .unwrap();
+        let times = [libc::timeval {
+            tv_sec: 0,
+            tv_usec: 0,
+        }; 2];
+        assert_eq!(unsafe { libc::utimes(c_path.as_ptr(), times.as_ptr()) }, 0);
+    }
+
+    #[test]
+    fn test_kvp_pool_file_names() {
+        assert_eq!(KvpPool::External.file_name(), ".kvp_pool_0");
+        assert_eq!(KvpPool::Guest.file_name(), ".kvp_pool_1");
+        assert_eq!(KvpPool::Auto.file_name(), ".kvp_pool_2");
+        assert_eq!(KvpPool::AutoExternal.file_name(), ".kvp_pool_3");
+        assert_eq!(KvpPool::AutoInternal.file_name(), ".kvp_pool_4");
+    }
+
+    #[test]
+    fn test_kvp_pool_default_dir_and_path() {
+        assert_eq!(KvpPool::default_dir(), Path::new("/var/lib/hyperv"));
+        assert_eq!(
+            KvpPool::Guest.default_path(),
+            Path::new("/var/lib/hyperv/.kvp_pool_1")
+        );
+    }
+
+    #[test]
+    fn test_default_path_is_used() {
+        let store = KvpPoolStore::new(KvpPool::Guest, PoolMode::Safe).unwrap();
+        assert_eq!(store.path(), KvpPool::Guest.default_path());
+    }
+
+    #[test]
+    fn test_new_in_derives_path_from_pool() {
+        let dir = TempDir::new().unwrap();
+        let store =
+            KvpPoolStore::new_in(KvpPool::Guest, dir.path(), PoolMode::Safe)
+                .unwrap();
+        assert_eq!(store.path(), dir.path().join(KvpPool::Guest.file_name()));
+    }
+
+    #[test]
+    fn test_pool_getter() {
+        let store =
+            KvpPoolStore::new(KvpPool::External, PoolMode::Safe).unwrap();
+        assert_eq!(store.pool(), KvpPool::External);
+    }
+
+    #[test]
+    fn test_mode_getter() {
+        let dir = TempDir::new().unwrap();
+        let store_safe = safe_store(dir.path());
+        assert_eq!(store_safe.mode(), PoolMode::Safe);
+
+        let dir2 = TempDir::new().unwrap();
+        let store_unsafe = unsafe_store(dir2.path());
+        assert_eq!(store_unsafe.mode(), PoolMode::Unsafe);
+    }
+
+    #[test]
+    fn test_max_key_size() {
+        let dir = TempDir::new().unwrap();
+        assert_eq!(safe_store(dir.path()).max_key_size(), 254);
+
+        let dir2 = TempDir::new().unwrap();
+        assert_eq!(unsafe_store(dir2.path()).max_key_size(), 512);
+    }
+
+    #[test]
+    fn test_max_value_size() {
+        let dir = TempDir::new().unwrap();
+        assert_eq!(safe_store(dir.path()).max_value_size(), 1022);
+
+        let dir2 = TempDir::new().unwrap();
+        assert_eq!(unsafe_store(dir2.path()).max_value_size(), 2048);
+    }
+
+    #[test]
+    fn test_insert_rejects_empty_key() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        let err = store.insert("", "value").unwrap_err();
+        assert!(matches!(err, KvpError::EmptyKey));
+    }
+
+    #[test]
+    fn test_insert_rejects_null_in_key() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        let err = store.insert("bad\0key", "value").unwrap_err();
+        assert!(matches!(err, KvpError::KeyContainsNull));
+    }
+
+    #[test]
+    fn test_safe_limits_key_254_pass_255_fail() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        let ok_key = "k".repeat(254);
+        store.insert(&ok_key, "v").unwrap();
+
+        let bad_key = "k".repeat(255);
+        let err = store.insert(&bad_key, "v").unwrap_err();
+        assert!(matches!(err, KvpError::KeyTooLarge { .. }));
+    }
+
+    #[test]
+    fn test_safe_limits_value_1022_pass_1023_fail() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        let ok_val = "v".repeat(1022);
+        store.insert("k", &ok_val).unwrap();
+
+        let bad_val = "v".repeat(1023);
+        let err = store.insert("k2", &bad_val).unwrap_err();
+        assert!(matches!(err, KvpError::ValueTooLarge { .. }));
+    }
+
+    #[test]
+    fn test_unsafe_limits_key_512_pass_513_fail() {
+        let dir = TempDir::new().unwrap();
+        let store = unsafe_store(dir.path());
+
+        let ok_key = "k".repeat(512);
+        store.insert(&ok_key, "v").unwrap();
+
+        let bad_key = "k".repeat(513);
+        let err = store.insert(&bad_key, "v").unwrap_err();
+        assert!(matches!(err, KvpError::KeyTooLarge { .. }));
+    }
+
+    #[test]
+    fn test_unsafe_limits_value_2048_pass_2049_fail() {
+        let dir = TempDir::new().unwrap();
+        let store = unsafe_store(dir.path());
+
+        let ok_val = "v".repeat(2048);
+        store.insert("k", &ok_val).unwrap();
+
+        let bad_val = "v".repeat(2049);
+        let err = store.insert("k2", &bad_val).unwrap_err();
+        assert!(matches!(err, KvpError::ValueTooLarge { .. }));
+    }
+
+    #[test]
+    fn test_read_accepts_wire_max_key_in_safe_mode() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        let store_unsafe = unsafe_store(dir.path());
+        let long_key = "k".repeat(300);
+        store_unsafe.insert(&long_key, "val").unwrap();
+
+        assert_eq!(store.read(&long_key).unwrap(), Some("val".to_string()));
+
+        let too_long = "k".repeat(513);
+        let err = store.read(&too_long).unwrap_err();
+        assert!(matches!(err, KvpError::KeyTooLarge { .. }));
+    }
+
+    #[test]
+    fn test_append_validates_sizes() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        let big_key = "k".repeat(255);
+        let err = store.append(&big_key, "v").unwrap_err();
+        assert!(matches!(err, KvpError::KeyTooLarge { .. }));
+
+        let big_val = "v".repeat(1023);
+        let err = store.append("k", &big_val).unwrap_err();
+        assert!(matches!(err, KvpError::ValueTooLarge { .. }));
+    }
+
+    #[test]
+    fn test_insert_overwrites_existing_key() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        store.insert("key", "v1").unwrap();
+        store.insert("key", "v2").unwrap();
+        store.insert("other", "v3").unwrap();
+
+        let entries = store.entries().unwrap();
+        assert_eq!(entries.len(), 2);
+        assert_eq!(entries.get("key"), Some(&"v2".to_string()));
+        assert_eq!(entries.get("other"), Some(&"v3".to_string()));
+    }
+
+    #[test]
+    fn test_insert_file_size_unchanged_on_overwrite() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        store.insert("key", "original").unwrap();
+        let size_before = std::fs::metadata(store.path()).unwrap().len();
+
+        store.insert("key", "updated").unwrap();
+        let size_after = std::fs::metadata(store.path()).unwrap().len();
+
+        assert_eq!(size_before, size_after);
+        assert_eq!(store.read("key").unwrap(), Some("updated".to_string()));
+    }
+
+    #[test]
+    fn test_insert_file_size_grows_on_new_key() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        store.insert("k1", "v1").unwrap();
+        let size_before = std::fs::metadata(store.path()).unwrap().len();
+
+        store.insert("k2", "v2").unwrap();
+        let size_after = std::fs::metadata(store.path()).unwrap().len();
+
+        assert_eq!(size_after - size_before, RECORD_SIZE as u64);
+    }
+
+    #[test]
+    fn test_insert_preserves_other_records() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        for i in 0..5 {
+            store.insert(&format!("k{i}"), &format!("v{i}")).unwrap();
+        }
+
+        let raw_before = std::fs::read(store.path()).unwrap();
+        store.insert("k2", "CHANGED").unwrap();
+        let raw_after = std::fs::read(store.path()).unwrap();
+
+        assert_eq!(raw_before.len(), raw_after.len());
+        for i in 0..5 {
+            if i == 2 {
+                continue;
+            }
+            let start = i * RECORD_SIZE;
+            let end = start + RECORD_SIZE;
+            assert_eq!(
+                &raw_before[start..end],
+                &raw_after[start..end],
+                "record {i} was modified unexpectedly"
+            );
+        }
+    }
+
+    #[test]
+    fn test_insert_after_append_duplicates_collapses_to_one() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        store.append("k", "v1").unwrap();
+        store.append("k", "v2").unwrap();
+        store.append("k", "v3").unwrap();
+        assert_eq!(store.len().unwrap(), 3);
+
+        store.insert("k", "updated").unwrap();
+        assert_eq!(store.read("k").unwrap(), Some("updated".to_string()));
+        assert_eq!(store.len().unwrap(), 1);
+
+        let raw = store.dump().unwrap();
+        assert_eq!(raw.len(), 1);
+        assert_eq!(raw[0], ("k".to_string(), "updated".to_string()));
+    }
+
+    #[test]
+    fn test_unique_key_cap_allows_1024_then_rejects_1025th() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        prepopulate(&store, MAX_UNIQUE_KEYS);
+        assert_eq!(store.len().unwrap(), MAX_UNIQUE_KEYS);
+
+        let err = store.insert("overflow", "v").unwrap_err();
+        assert!(matches!(err, KvpError::MaxUniqueKeysExceeded { .. }));
+    }
+
+    #[test]
+    fn test_unique_key_cap_allows_overwrite_at_limit() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        prepopulate(&store, MAX_UNIQUE_KEYS);
+
+        store.insert("k0", "updated").unwrap();
+        assert_eq!(store.read("k0").unwrap(), Some("updated".to_string()));
+    }
+
+    #[test]
+    fn test_insert_cap_uses_unique_keys_not_record_count() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        prepopulate(&store, MAX_UNIQUE_KEYS - 1);
+
+        store.append("k0", "dup").unwrap();
+        assert_eq!(store.len().unwrap(), MAX_UNIQUE_KEYS);
+
+        store.insert("new_unique", "v").unwrap();
+        assert_eq!(store.read("new_unique").unwrap(), Some("v".to_string()));
+    }
+
+    #[test]
+    fn test_unique_key_cap_allows_new_key_after_delete() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        prepopulate(&store, MAX_UNIQUE_KEYS);
+
+        assert!(store.delete("k0").unwrap());
+        store.insert("new-key", "v").unwrap();
+    }
+
+    #[test]
+    fn test_append_no_key_cap() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        prepopulate(&store, MAX_UNIQUE_KEYS);
+
+        store.append("extra", "val").unwrap();
+        assert_eq!(store.len().unwrap(), MAX_UNIQUE_KEYS + 1);
+    }
+
+    #[test]
+    fn test_append_then_read() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        store.append("diag_001", "event_data").unwrap();
+        assert_eq!(
+            store.read("diag_001").unwrap(),
+            Some("event_data".to_string())
+        );
+    }
+
+    #[test]
+    fn test_append_duplicate_key_semantics() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        store.append("X", "first").unwrap();
+        store.append("X", "second").unwrap();
+
+        assert_eq!(store.read("X").unwrap(), Some("second".to_string()));
+
+        let entries = store.entries().unwrap();
+        assert_eq!(entries.get("X"), Some(&"second".to_string()));
+    }
+
+    #[test]
+    fn test_read_missing_file_returns_none() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        assert_eq!(store.read("anything").unwrap(), None);
+    }
+
+    #[test]
+    fn test_read_rejects_empty_key() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        let err = store.read("").unwrap_err();
+        assert!(matches!(err, KvpError::EmptyKey));
+    }
+
+    #[test]
+    fn test_read_rejects_null_in_key() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        let err = store.read("bad\0key").unwrap_err();
+        assert!(matches!(err, KvpError::KeyContainsNull));
+    }
+
+    #[test]
+    fn test_delete_removes_key() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        store.insert("key", "v1").unwrap();
+        store.insert("other", "v2").unwrap();
+
+        assert!(store.delete("key").unwrap());
+        assert_eq!(store.read("key").unwrap(), None);
+        assert_eq!(store.read("other").unwrap(), Some("v2".to_string()));
+    }
+
+    #[test]
+    fn test_delete_removes_all_duplicates() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        store.append("k", "v1").unwrap();
+        store.append("other", "x").unwrap();
+        store.append("k", "v2").unwrap();
+        store.append("k", "v3").unwrap();
+        assert_eq!(store.len().unwrap(), 4);
+
+        assert!(store.delete("k").unwrap());
+        assert_eq!(store.read("k").unwrap(), None);
+        assert_eq!(store.len().unwrap(), 1);
+        assert_eq!(store.read("other").unwrap(), Some("x".to_string()));
+    }
+
+    #[test]
+    fn test_delete_missing_file_returns_false() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        assert!(!store.delete("anything").unwrap());
+    }
+
+    #[test]
+    fn test_delete_rejects_empty_key() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        let err = store.delete("").unwrap_err();
+        assert!(matches!(err, KvpError::EmptyKey));
+    }
+
+    #[test]
+    fn test_delete_rejects_null_in_key() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        let err = store.delete("bad\0key").unwrap_err();
+        assert!(matches!(err, KvpError::KeyContainsNull));
+    }
+
+    #[test]
+    fn test_delete_nonexistent_key_returns_false() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.insert("k1", "v1").unwrap();
+
+        assert!(!store.delete("missing").unwrap());
+        assert_eq!(store.len().unwrap(), 1);
+    }
+
+    #[test]
+    fn test_clear_empties_store() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        store.insert("key", "value").unwrap();
+        store.clear().unwrap();
+        assert_eq!(store.read("key").unwrap(), None);
+    }
+
+    #[test]
+    fn test_clear_missing_file_is_ok() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        store.clear().unwrap();
+    }
+
+    #[test]
+    fn test_entries_missing_file_returns_empty() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        assert!(store.entries().unwrap().is_empty());
+    }
+
+    #[test]
+    fn test_dump_returns_all_records_with_duplicates() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        store.append("key", "val1").unwrap();
+        store.append("key", "val2").unwrap();
+        store.append("key", "val3").unwrap();
+
+        let entries = store.entries().unwrap();
+        assert_eq!(entries.get("key"), Some(&"val3".to_string()));
+
+        let raw = store.dump().unwrap();
+        assert_eq!(raw.len(), 3);
+        assert_eq!(
+            raw,
+            vec![
+                ("key".to_string(), "val1".to_string()),
+                ("key".to_string(), "val2".to_string()),
+                ("key".to_string(), "val3".to_string()),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_dump_missing_file_returns_empty() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        assert!(store.dump().unwrap().is_empty());
+    }
+
+    #[test]
+    fn test_dump_json_writes_json() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        store.insert("key1", "value1").unwrap();
+        store.insert("key2", "value2").unwrap();
+
+        let dump_file = NamedTempFile::new().unwrap();
+        store.dump_json(dump_file.path()).unwrap();
+
+        let contents = std::fs::read_to_string(dump_file.path()).unwrap();
+        let parsed: HashMap<String, String> =
+            serde_json::from_str(&contents).unwrap();
+        assert_eq!(parsed.len(), 2);
+        assert_eq!(parsed.get("key1"), Some(&"value1".to_string()));
+        assert_eq!(parsed.get("key2"), Some(&"value2".to_string()));
+    }
+
+    #[test]
+    fn test_dump_json_empty_store() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        File::create(store.path()).unwrap();
+
+        let dump_file = NamedTempFile::new().unwrap();
+        store.dump_json(dump_file.path()).unwrap();
+
+        let contents = std::fs::read_to_string(dump_file.path()).unwrap();
+        let parsed: HashMap<String, String> =
+            serde_json::from_str(&contents).unwrap();
+        assert!(parsed.is_empty());
+    }
+
+    #[test]
+    fn test_len_and_is_empty() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        assert!(store.is_empty().unwrap());
+        assert_eq!(store.len().unwrap(), 0);
+
+        store.insert("key", "value").unwrap();
+        assert!(!store.is_empty().unwrap());
+        assert_eq!(store.len().unwrap(), 1);
+
+        store.insert("key2", "value2").unwrap();
+        assert_eq!(store.len().unwrap(), 2);
+
+        store.insert("key", "updated").unwrap();
+        assert_eq!(store.len().unwrap(), 2);
+    }
+
+    #[test]
+    fn test_is_stale_and_is_stale_at_boot_helpers() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.insert("key", "value").unwrap();
+
+        assert!(!store.is_stale().unwrap());
+        assert!(store.is_stale_at_boot(i64::MAX).unwrap());
+        assert!(!store.is_stale_at_boot(0).unwrap());
+    }
+
+    #[test]
+    fn test_is_stale_missing_file_returns_false() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        assert!(!store.is_stale().unwrap());
+    }
+
+    #[test]
+    fn test_clear_if_stale_noop_when_fresh() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.insert("key", "value").unwrap();
+
+        store.clear_if_stale().unwrap();
+        assert_eq!(store.read("key").unwrap(), Some("value".to_string()));
+    }
+
+    #[test]
+    fn test_clear_if_stale_clears_stale_data() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.insert("key", "value").unwrap();
+        assert!(!store.is_empty().unwrap());
+
+        set_mtime_to_epoch(store.path());
+        assert!(store.is_stale().unwrap());
+
+        store.clear_if_stale().unwrap();
+        assert!(store.is_empty().unwrap());
+    }
+
+    #[test]
+    fn test_encode_decode_round_trip() {
+        let record = encode_record("hello", "world");
+        assert_eq!(record.len(), RECORD_SIZE);
+
+        let (k, v) = decode_record(&record).unwrap();
+        assert_eq!(k, "hello");
+        assert_eq!(v, "world");
+    }
+
+    #[test]
+    fn test_encode_decode_max_length_key() {
+        let key = "x".repeat(WIRE_MAX_KEY_BYTES);
+        let record = encode_record(&key, "val");
+        let (k, v) = decode_record(&record).unwrap();
+        assert_eq!(k, key);
+        assert_eq!(v, "val");
+    }
+
+    #[test]
+    fn test_encode_decode_empty_value() {
+        let record = encode_record("key", "");
+        let (k, v) = decode_record(&record).unwrap();
+        assert_eq!(k, "key");
+        assert_eq!(v, "");
+    }
+
+    #[test]
+    fn test_decode_wrong_size_errors() {
+        let err = decode_record(&[0u8; RECORD_SIZE + 1]).unwrap_err();
+        assert_eq!(err.kind(), ErrorKind::Other);
+    }
+
+    #[test]
+    fn test_iter_yields_all_records_in_order() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        let expected: Vec<(&str, &str)> =
+            vec![("a", "1"), ("b", "2"), ("c", "3"), ("d", "4"), ("e", "5")];
+        for (k, v) in &expected {
+            store.append(k, v).unwrap();
+        }
+
+        let iter = store.iter().unwrap();
+        assert_eq!(iter.record_count(), 5);
+
+        let records: Vec<(String, String)> = iter.map(|r| r.unwrap()).collect();
+        for (i, (k, v)) in records.iter().enumerate() {
+            assert_eq!(k, expected[i].0);
+            assert_eq!(v, expected[i].1);
+        }
+    }
+
+    #[test]
+    fn test_iter_empty_file() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        File::create(store.path()).unwrap();
+
+        let iter = store.iter().unwrap();
+        assert_eq!(iter.record_count(), 0);
+        assert_eq!(iter.count(), 0);
+    }
+
+    #[test]
+    fn test_iter_malformed_file_error() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        std::fs::write(store.path(), [0u8; RECORD_SIZE + 1]).unwrap();
+
+        let err = store.iter().unwrap_err();
+        assert!(matches!(err, KvpError::Io(_)));
+    }
+
+    #[test]
+    fn test_iter_drop_mid_iteration_releases_lock() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.insert("k1", "v1").unwrap();
+        store.insert("k2", "v2").unwrap();
+
+        {
+            let mut iter = store.iter().unwrap();
+            iter.next().unwrap().unwrap();
+        }
+
+        store.insert("k3", "v3").unwrap();
+        assert_eq!(store.len().unwrap(), 3);
+    }
+
+    #[test]
+    fn test_overwrite_adjacent_records_untouched() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.append("first", "val_first").unwrap();
+        store.append("middle", "val_middle").unwrap();
+        store.append("last", "val_last").unwrap();
+
+        let raw_before = std::fs::read(store.path()).unwrap();
+        let first_before = &raw_before[..RECORD_SIZE];
+        let last_before = &raw_before[2 * RECORD_SIZE..3 * RECORD_SIZE];
+
+        {
+            let mut iter = store.iter_mut().unwrap();
+            iter.next().unwrap().unwrap();
+            let (k, _) = iter.next().unwrap().unwrap();
+            assert_eq!(k, "middle");
+            iter.overwrite_current_value("UPDATED").unwrap();
+            iter.flush().unwrap();
+        }
+
+        let raw_after = std::fs::read(store.path()).unwrap();
+        assert_eq!(&raw_after[..RECORD_SIZE], first_before);
+        assert_eq!(&raw_after[2 * RECORD_SIZE..3 * RECORD_SIZE], last_before);
+
+        assert_eq!(store.read("middle").unwrap(), Some("UPDATED".to_string()));
+    }
+
+    #[test]
+    fn test_overwrite_multiple_in_one_pass() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.append("a", "old_a").unwrap();
+        store.append("b", "old_b").unwrap();
+        store.append("c", "old_c").unwrap();
+
+        {
+            let mut iter = store.iter_mut().unwrap();
+            let mut idx = 0;
+            while let Some(record) = iter.next() {
+                record.unwrap();
+                if idx == 0 || idx == 2 {
+                    iter.overwrite_current_value("new").unwrap();
+                }
+                idx += 1;
+            }
+            iter.flush().unwrap();
+        }
+
+        assert_eq!(store.read("a").unwrap(), Some("new".to_string()));
+        assert_eq!(store.read("b").unwrap(), Some("old_b".to_string()));
+        assert_eq!(store.read("c").unwrap(), Some("new".to_string()));
+    }
+
+    #[test]
+    fn test_overwrite_before_next_errors() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.append("k", "v").unwrap();
+
+        let mut iter = store.iter_mut().unwrap();
+        let err = iter.overwrite_current_value("new").unwrap_err();
+        assert_eq!(err.kind(), ErrorKind::InvalidInput);
+    }
+
+    #[test]
+    fn test_remove_before_next_errors() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.append("k", "v").unwrap();
+
+        let mut iter = store.iter_mut().unwrap();
+        let err = iter.remove_current().unwrap_err();
+        assert_eq!(err.kind(), ErrorKind::InvalidInput);
+    }
+
+    #[test]
+    fn test_remove_swap_and_continue() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.append("A", "1").unwrap();
+        store.append("B", "2").unwrap();
+        store.append("C", "3").unwrap();
+        store.append("D", "4").unwrap();
+
+        let mut collected = Vec::new();
+        {
+            let mut iter = store.iter_mut().unwrap();
+            while let Some(record) = iter.next() {
+                let (k, v) = record.unwrap();
+                if k == "B" {
+                    iter.remove_current().unwrap();
+                } else {
+                    collected.push((k, v));
+                }
+            }
+            iter.flush().unwrap();
+        }
+
+        assert_eq!(collected.len(), 3);
+        assert!(collected.iter().any(|(k, _)| k == "A"));
+        assert!(collected.iter().any(|(k, _)| k == "C"));
+        assert!(collected.iter().any(|(k, _)| k == "D"));
+
+        assert_eq!(store.len().unwrap(), 3);
+        assert_eq!(store.read("B").unwrap(), None);
+    }
+
+    #[test]
+    fn test_remove_last_record_no_swap() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.append("a", "1").unwrap();
+        store.append("b", "2").unwrap();
+        store.append("c", "3").unwrap();
+
+        {
+            let mut iter = store.iter_mut().unwrap();
+            iter.next().unwrap().unwrap();
+            iter.next().unwrap().unwrap();
+            iter.next().unwrap().unwrap();
+            iter.remove_current().unwrap();
+            iter.flush().unwrap();
+        }
+
+        assert_eq!(store.len().unwrap(), 2);
+        assert_eq!(store.read("c").unwrap(), None);
+        assert_eq!(store.read("a").unwrap(), Some("1".to_string()));
+        assert_eq!(store.read("b").unwrap(), Some("2".to_string()));
+    }
+
+    #[test]
+    fn test_remove_only_record() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.append("solo", "val").unwrap();
+
+        {
+            let mut iter = store.iter_mut().unwrap();
+            iter.next().unwrap().unwrap();
+            iter.remove_current().unwrap();
+            iter.flush().unwrap();
+        }
+
+        assert!(store.is_empty().unwrap());
+    }
+
+    #[test]
+    fn test_file_integrity_after_mixed_ops() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+
+        for i in 0..10 {
+            store.insert(&format!("u{i}"), &format!("uv{i}")).unwrap();
+        }
+        for i in 0..5 {
+            store.append(&format!("a{i}"), &format!("av{i}")).unwrap();
+        }
+        store.delete("u3").unwrap();
+        store.delete("u7").unwrap();
+        store.insert("u0", "overwritten").unwrap();
+
+        let raw = std::fs::read(store.path()).unwrap();
+        assert_eq!(raw.len() % RECORD_SIZE, 0);
+
+        let record_count = raw.len() / RECORD_SIZE;
+        for i in 0..record_count {
+            let start = i * RECORD_SIZE;
+            let end = start + RECORD_SIZE;
+            decode_record(&raw[start..end]).unwrap_or_else(|e| {
+                panic!("record {i} failed to decode: {e:?}")
+            });
+        }
+    }
+
+    #[test]
+    fn test_kvp_error_display() {
+        let cases: Vec<(KvpError, &str)> = vec![
+            (KvpError::EmptyKey, "KVP key must not be empty"),
+            (
+                KvpError::KeyTooLarge {
+                    max: 254,
+                    actual: 300,
+                },
+                "KVP key length (300) exceeds maximum (254)",
+            ),
+            (
+                KvpError::ValueTooLarge {
+                    max: 1022,
+                    actual: 2000,
+                },
+                "KVP value length (2000) exceeds maximum (1022)",
+            ),
+            (
+                KvpError::MaxUniqueKeysExceeded { max: 1024 },
+                "KVP unique key count exceeded maximum (1024)",
+            ),
+            (
+                KvpError::KeyContainsNull,
+                "KVP key must not contain null bytes",
+            ),
+            (
+                KvpError::Io(io::Error::new(ErrorKind::NotFound, "gone")),
+                "gone",
+            ),
+        ];
+        for (err, expected) in cases {
+            assert_eq!(err.to_string(), expected);
+        }
+    }
+
+    #[test]
+    fn test_kvp_error_source() {
+        let io_err = KvpError::Io(io::Error::other("kvp-error-source"));
+        assert!(io_err.source().is_some());
+
+        assert!(KvpError::EmptyKey.source().is_none());
+        assert!(KvpError::KeyContainsNull.source().is_none());
+    }
+
+    #[test]
+    fn test_kvp_error_from_io() {
+        let io_err = io::Error::new(ErrorKind::PermissionDenied, "denied");
+        let kvp_err: KvpError = io_err.into();
+        assert!(matches!(kvp_err, KvpError::Io(_)));
+        assert_eq!(kvp_err.to_string(), "denied");
+    }
+
+    #[test]
+    fn test_concurrent_inserts_to_different_keys() {
+        let dir = TempDir::new().unwrap();
+        let dir_path = dir.path().to_path_buf();
+        let threads: Vec<_> = (0..8)
+            .map(|t| {
+                let d = dir_path.clone();
+                std::thread::spawn(move || {
+                    let store =
+                        KvpPoolStore::new_in(KvpPool::Guest, d, PoolMode::Safe)
+                            .unwrap();
+                    for i in 0..10 {
+                        let key = format!("t{t}_k{i}");
+                        store.insert(&key, &format!("val_{t}_{i}")).unwrap();
+                    }
+                })
+            })
+            .collect();
+
+        for t in threads {
+            t.join().unwrap();
+        }
+
+        let store = safe_store(dir.path());
+        let entries = store.entries().unwrap();
+        assert_eq!(entries.len(), 80);
+        for t in 0..8 {
+            for i in 0..10 {
+                let key = format!("t{t}_k{i}");
+                assert_eq!(
+                    entries.get(&key),
+                    Some(&format!("val_{t}_{i}")),
+                    "missing or wrong value for {key}"
+                );
+            }
+        }
+    }
+
+    #[test]
+    fn test_concurrent_inserts_to_same_key() {
+        let dir = TempDir::new().unwrap();
+        let dir_path = dir.path().to_path_buf();
+        let threads: Vec<_> = (0..8)
+            .map(|t| {
+                let d = dir_path.clone();
+                std::thread::spawn(move || {
+                    let store =
+                        KvpPoolStore::new_in(KvpPool::Guest, d, PoolMode::Safe)
+                            .unwrap();
+                    for i in 0..10 {
+                        store
+                            .insert("shared_key", &format!("t{t}_v{i}"))
+                            .unwrap();
+                    }
+                })
+            })
+            .collect();
+
+        for t in threads {
+            t.join().unwrap();
+        }
+
+        let store = safe_store(dir.path());
+        assert_eq!(store.len().unwrap(), 1);
+        let val = store.read("shared_key").unwrap().unwrap();
+        assert!(
+            val.starts_with('t') && val.contains("_v"),
+            "unexpected value format: {val}"
+        );
+
+        let file_len = std::fs::metadata(store.path()).unwrap().len() as usize;
+        assert_eq!(file_len, RECORD_SIZE);
+    }
+
+    #[test]
+    fn test_concurrent_appends() {
+        let dir = TempDir::new().unwrap();
+        let dir_path = dir.path().to_path_buf();
+
+        let threads: Vec<_> = (0..8)
+            .map(|t| {
+                let d = dir_path.clone();
+                std::thread::spawn(move || {
+                    let store =
+                        KvpPoolStore::new_in(KvpPool::Guest, d, PoolMode::Safe)
+                            .unwrap();
+                    for i in 0..10 {
+                        store
+                            .append(
+                                &format!("t{t}_d{i}"),
+                                &format!("data_{t}_{i}"),
+                            )
+                            .unwrap();
+                    }
+                })
+            })
+            .collect();
+
+        for t in threads {
+            t.join().unwrap();
+        }
+
+        let store = safe_store(dir.path());
+        assert_eq!(store.len().unwrap(), 80);
+
+        let entries = store.entries().unwrap();
+        for t in 0..8 {
+            for i in 0..10 {
+                let key = format!("t{t}_d{i}");
+                assert!(entries.contains_key(&key), "missing key {key}");
+            }
+        }
+    }
+
+    #[test]
+    fn test_concurrent_writers_at_key_cap() {
+        let dir = TempDir::new().unwrap();
+        let dir_path = dir.path().to_path_buf();
+
+        let store = safe_store(dir.path());
+        prepopulate(&store, MAX_UNIQUE_KEYS - 4);
+
+        let threads: Vec<_> = (0..4)
+            .map(|t| {
+                let d = dir_path.clone();
+                std::thread::spawn(move || {
+                    let store =
+                        KvpPoolStore::new_in(KvpPool::Guest, d, PoolMode::Safe)
+                            .unwrap();
+                    store.insert(&format!("new{t}"), "v")
+                })
+            })
+            .collect();
+
+        let results: Vec<_> =
+            threads.into_iter().map(|t| t.join().unwrap()).collect();
+
+        let successes = results.iter().filter(|r| r.is_ok()).count();
+        let cap_errors = results
+            .iter()
+            .filter(|r| {
+                matches!(r, Err(KvpError::MaxUniqueKeysExceeded { .. }))
+            })
+            .count();
+
+        assert_eq!(successes, 4);
+        assert_eq!(cap_errors, 0);
+        assert_eq!(store.len().unwrap(), MAX_UNIQUE_KEYS);
+
+        let err = store.insert("one_too_many", "v").unwrap_err();
+        assert!(matches!(err, KvpError::MaxUniqueKeysExceeded { .. }));
+    }
+
+    #[test]
+    fn test_concurrent_readers_only() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        for i in 0..10 {
+            store.insert(&format!("k{i}"), &format!("v{i}")).unwrap();
+        }
+
+        let dir_path = dir.path().to_path_buf();
+        let threads: Vec<_> = (0..8)
+            .map(|_| {
+                let d = dir_path.clone();
+                std::thread::spawn(move || {
+                    let store =
+                        KvpPoolStore::new_in(KvpPool::Guest, d, PoolMode::Safe)
+                            .unwrap();
+                    for _ in 0..20 {
+                        let entries = store.entries().unwrap();
+                        assert_eq!(entries.len(), 10);
+                        for i in 0..10 {
+                            assert_eq!(
+                                store.read(&format!("k{i}")).unwrap(),
+                                Some(format!("v{i}"))
+                            );
+                        }
+                    }
+                })
+            })
+            .collect();
+
+        for t in threads {
+            t.join().unwrap();
+        }
+    }
+
+    #[test]
+    fn test_concurrent_readers_and_writers() {
+        let dir = TempDir::new().unwrap();
+        let dir_path = dir.path().to_path_buf();
+
+        let store = safe_store(dir.path());
+        for i in 0..10 {
+            store.insert(&format!("k{i}"), &format!("v{i}")).unwrap();
+        }
+
+        let writer_threads: Vec<_> = (0..4)
+            .map(|t| {
+                let d = dir_path.clone();
+                std::thread::spawn(move || {
+                    let store =
+                        KvpPoolStore::new_in(KvpPool::Guest, d, PoolMode::Safe)
+                            .unwrap();
+                    for round in 0..5 {
+                        let key = format!("k{t}");
+                        store.insert(&key, &format!("w{t}_r{round}")).unwrap();
+                    }
+                })
+            })
+            .collect();
+
+        let reader_threads: Vec<_> = (0..4)
+            .map(|_| {
+                let d = dir_path.clone();
+                std::thread::spawn(move || {
+                    let store =
+                        KvpPoolStore::new_in(KvpPool::Guest, d, PoolMode::Safe)
+                            .unwrap();
+                    for _ in 0..20 {
+                        let entries = store.entries().unwrap();
+                        assert!(entries.len() <= 10);
+                        for (k, v) in &entries {
+                            assert!(!k.is_empty());
+                            assert!(!v.is_empty());
+                        }
+                    }
+                })
+            })
+            .collect();
+
+        for t in writer_threads {
+            t.join().unwrap();
+        }
+        for t in reader_threads {
+            t.join().unwrap();
+        }
+
+        let entries = store.entries().unwrap();
+        assert_eq!(entries.len(), 10);
+    }
+
+    #[test]
+    fn test_concurrent_mixed_insert_and_append() {
+        let dir = TempDir::new().unwrap();
+        let dir_path = dir.path().to_path_buf();
+
+        let store = safe_store(dir.path());
+        for i in 0..4 {
+            store.insert(&format!("fixed{i}"), "init").unwrap();
+        }
+
+        let insert_threads: Vec<_> = (0..4)
+            .map(|t| {
+                let d = dir_path.clone();
+                std::thread::spawn(move || {
+                    let store =
+                        KvpPoolStore::new_in(KvpPool::Guest, d, PoolMode::Safe)
+                            .unwrap();
+                    for round in 0..5 {
+                        store
+                            .insert(&format!("fixed{t}"), &format!("r{round}"))
+                            .unwrap();
+                    }
+                })
+            })
+            .collect();
+
+        let append_threads: Vec<_> = (0..4)
+            .map(|t| {
+                let d = dir_path.clone();
+                std::thread::spawn(move || {
+                    let store =
+                        KvpPoolStore::new_in(KvpPool::Guest, d, PoolMode::Safe)
+                            .unwrap();
+                    for i in 0..5 {
+                        store
+                            .append(
+                                &format!("diag_t{t}_i{i}"),
+                                &format!("d{i}"),
+                            )
+                            .unwrap();
+                    }
+                })
+            })
+            .collect();
+
+        for t in insert_threads {
+            t.join().unwrap();
+        }
+        for t in append_threads {
+            t.join().unwrap();
+        }
+
+        let entries = store.entries().unwrap();
+        assert_eq!(entries.len(), 24);
+        for t in 0..4 {
+            assert!(entries.contains_key(&format!("fixed{t}")));
+        }
+    }
+
+    #[test]
+    fn test_read_permission_denied() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.insert("k", "v").unwrap();
+
+        lock_dir(dir.path());
+        let err = store.read("k").unwrap_err();
+        unlock_dir(dir.path());
+        assert!(matches!(err, KvpError::Io(_)));
+    }
+
+    #[test]
+    fn test_delete_permission_denied() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.insert("k", "v").unwrap();
+
+        lock_dir(dir.path());
+        let err = store.delete("k").unwrap_err();
+        unlock_dir(dir.path());
+        assert!(matches!(err, KvpError::Io(_)));
+    }
+
+    #[test]
+    fn test_clear_permission_denied() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.insert("k", "v").unwrap();
+
+        lock_dir(dir.path());
+        let err = store.clear().unwrap_err();
+        unlock_dir(dir.path());
+        assert!(matches!(err, KvpError::Io(_)));
+    }
+
+    #[test]
+    fn test_entries_permission_denied() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.insert("k", "v").unwrap();
+
+        lock_dir(dir.path());
+        let err = store.entries().unwrap_err();
+        unlock_dir(dir.path());
+        assert!(matches!(err, KvpError::Io(_)));
+    }
+
+    #[test]
+    fn test_dump_permission_denied() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.insert("k", "v").unwrap();
+
+        lock_dir(dir.path());
+        let err = store.dump().unwrap_err();
+        unlock_dir(dir.path());
+        assert!(matches!(err, KvpError::Io(_)));
+    }
+
+    #[test]
+    fn test_len_permission_denied() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.insert("k", "v").unwrap();
+
+        lock_dir(dir.path());
+        let err = store.len().unwrap_err();
+        unlock_dir(dir.path());
+        assert!(matches!(err, KvpError::Io(_)));
+    }
+
+    #[test]
+    fn test_is_stale_permission_denied() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.insert("k", "v").unwrap();
+
+        lock_dir(dir.path());
+        let err = store.is_stale().unwrap_err();
+        unlock_dir(dir.path());
+        assert!(matches!(err, KvpError::Io(_)));
+    }
+
+    #[test]
+    fn test_is_stale_at_boot_missing_file() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        assert!(!store.is_stale_at_boot(i64::MAX).unwrap());
+    }
+
+    #[test]
+    fn test_is_stale_at_boot_permission_denied() {
+        let dir = TempDir::new().unwrap();
+        let store = safe_store(dir.path());
+        store.insert("k", "v").unwrap();
+
+        lock_dir(dir.path());
+        let err = store.is_stale_at_boot(i64::MAX).unwrap_err();
+        unlock_dir(dir.path());
+        assert!(matches!(err, KvpError::Io(_)));
+    }
+}
